chore(refactor): extract /api/memories into routes/memories.ts
Fourth bucket of the index.ts route extraction. Tiny — two endpoints:
  GET  /api/memories
  POST /api/memories

Pattern unchanged from sessions/artifacts/spaces. No behavioural change.

Cumulative: index.ts now down from 2218 (pre-audit) → 1444 (-35%).
