+++ server/src/index.ts
@@ -28,6 +28,7 @@ import { makeRouteCtx } from "./http-utils.js";
 import { tryHandleSessionRoute } from "./routes/sessions.js";
 import { tryHandleArtifactRoute } from "./routes/artifacts.js";
 import { tryHandleSpaceRoute } from "./routes/spaces.js";
+import { tryHandleMemoryRoute } from "./routes/memories.js";
 import type { UiCommand } from "../../shared/types.js";
 import {
   scanExistingArtifacts,
@@ -676,6 +677,9 @@ async function handleHttpRequest(req: IncomingMessage, res: ServerResponse) {
     spaceService, broadcastUiEvent,
   })) return;
 
+  // /api/memories
+  if (await tryHandleMemoryRoute(req, res, url, ctx, { memoryProvider })) return;
+
   // GET /api/resolve-path?url=...  — resolve a serving URL to a filesystem path
   if (url.startsWith("/api/resolve-path")) {
     const params = new URL(url, "http://localhost").searchParams;
@@ -757,48 +761,6 @@ async function handleHttpRequest(req: IncomingMessage, res: ServerResponse) {
       sendError(err, 500);
     }
     return;
-  }
-
-  // GET /api/memories — list memories, optionally scoped to a space.
-  // Local-origin only: memory contents are private user notes. Strip the
-  // query string before path-matching (same trap the events route had —
-  // `$`-anchored regex would silently reject `?space_id=…`).
-  {
-    const memoriesPath = url.split("?")[0];
-    if (memoriesPath === "/api/memories" && req.method === "GET") {
-      if (rejectIfNonLocalOrigin()) return;
-      const parsed = new URL(req.url ?? "/", "http://localhost");
-      const spaceId = parsed.searchParams.get("space_id");
-      try {
-        const memories = await memoryProvider.list(spaceId ?? undefined);
-        sendJson(memories);
-      } catch (err) {
-        sendError(err, 500);
-      }
-      return;
-    }
-    // POST /api/memories — user-authored memory. Mirrors the MCP `remember`
-    // tool: empty content rejected, exact-content dedupe, space optional.
-    if (memoriesPath === "/api/memories" && req.method === "POST") {
-      if (rejectIfNonLocalOrigin()) return;
-      try {
-        const body = await readJsonBody();
-        const content = typeof body.content === "string" ? body.content.trim() : "";
-        if (!content) {
-          sendJson({ error: "content is required" }, 400);
-          return;
-        }
-        const space_id = typeof body.space_id === "string" && body.space_id ? body.space_id : undefined;
-        const tags = Array.isArray(body.tags)
-          ? body.tags.filter((t): t is string => typeof t === "string" && t.length > 0)
-          : undefined;
-        const memory = await memoryProvider.remember({ content, space_id, tags });
-        sendJson(memory, 201);
-      } catch (err) {
-        sendError(err);
-      }
-      return;
-    }
   }
 
   // GET /api/ui/events — SSE stream for UI commands.
