+++ server/src/routes/memories.ts
@@ -1,0 +1,65 @@
+// /api/memories — extracted from index.ts. Two endpoints, both
+// local-origin only (memory contents are private user notes).
+
+import type { IncomingMessage, ServerResponse } from "node:http";
+import type { MemoryProvider } from "../memory-store.js";
+import type { RouteCtx } from "../http-utils.js";
+
+export interface MemoryRouteDeps {
+  memoryProvider: MemoryProvider;
+}
+
+export async function tryHandleMemoryRoute(
+  req: IncomingMessage,
+  res: ServerResponse,
+  url: string,
+  ctx: RouteCtx,
+  deps: MemoryRouteDeps,
+): Promise<boolean> {
+  const { sendJson, sendError, readJsonBody, rejectIfNonLocalOrigin } = ctx;
+  const { memoryProvider } = deps;
+
+  // GET /api/memories — list memories, optionally scoped to a space.
+  // Strip the query string before path-matching (same trap the events
+  // route had — `$`-anchored regex would silently reject `?space_id=…`).
+  const memoriesPath = url.split("?")[0];
+  if (memoriesPath !== "/api/memories") return false;
+
+  if (req.method === "GET") {
+    if (rejectIfNonLocalOrigin()) return true;
+    const parsed = new URL(req.url ?? "/", "http://localhost");
+    const spaceId = parsed.searchParams.get("space_id");
+    try {
+      const memories = await memoryProvider.list(spaceId ?? undefined);
+      sendJson(memories);
+    } catch (err) {
+      sendError(err, 500);
+    }
+    return true;
+  }
+
+  // POST /api/memories — user-authored memory. Mirrors the MCP `remember`
+  // tool: empty content rejected, exact-content dedupe, space optional.
+  if (req.method === "POST") {
+    if (rejectIfNonLocalOrigin()) return true;
+    try {
+      const body = await readJsonBody();
+      const content = typeof body.content === "string" ? body.content.trim() : "";
+      if (!content) {
+        sendJson({ error: "content is required" }, 400);
+        return true;
+      }
+      const space_id = typeof body.space_id === "string" && body.space_id ? body.space_id : undefined;
+      const tags = Array.isArray(body.tags)
+        ? body.tags.filter((t): t is string => typeof t === "string" && t.length > 0)
+        : undefined;
+      const memory = await memoryProvider.remember({ content, space_id, tags });
+      sendJson(memory, 201);
+    } catch (err) {
+      sendError(err);
+    }
+    return true;
+  }
+
+  return false;
+}
